+++ utils.h
@@ -6,7 +6,7 @@
 
 DWORD lendian(BYTE *, uint8_t);
 char *sign8x(BYTE);
-uint8_t valid_addr(PESTRUCT *, DWORD);
+int valid_addr(PESTRUCT *, DWORD);
 DWORD addr_to_offset(PESTRUCT *, DWORD);
 char *get_instr(FILE *, PESTRUCT *, DWORD);
 DWORD parse_addr(char *);
